refactor: apply coding standards to schemaorg files (#4806)
