+++ pages/[entityListType]/[...params].tsx
@@ -30,9 +30,9 @@ import { useRouter } from "next/router";
 import { ParsedUrlQuery } from "querystring";
 import { JSX } from "react";
 import { EntityGuard } from "../../app/components/Detail/components/EntityGuard/entityGuard";
-import { JsonLd } from "../../app/components/Detail/components/JsonLd/jsonLd";
 import { buildHcaProjectJsonLd } from "../../app/utils/schemaOrg/hcaProjectDataset";
 import { readFile } from "../../app/utils/tsvParser";
+import { JsonLd } from "../../app/views/EntityDetailView/components/JsonLd/jsonLd";
 import { ROUTES } from "../../site-config/anvil-cmg/dev/export/routes";
 
 import { DatasetsResponse } from "../../app/apis/azul/anvil-cmg/common/responses";
@@ -40,7 +40,7 @@ import {
   getConsentGroup,
   isNRESOrUnrestrictedAccess,
 } from "../../app/apis/azul/anvil-cmg/common/transformers";
-import { ProjectsResponse } from "../../app/apis/azul/hca-dcp/common/responses";
+import type { ProjectsResponse } from "../../app/apis/azul/hca-dcp/common/responses";
 import { isProductionEnvironment } from "../../app/config/utils";
 
 const setOfProcessedIds = new Set<string>();
@@ -93,24 +93,6 @@ const EntityDetailPage = (props: EntityDetailPageProps): JSX.Element => {
     </>
   );
 };
-
-/**
- * Renders the HCA project JSON-LD when the page is a project detail route with
- * data and a browser URL available. Returns null otherwise.
- * @param props - Entity detail page props.
- * @returns JsonLd element, or null when the page can't be described.
- */
-function renderHcaProjectJsonLd(
-  props: EntityDetailPageProps
-): JSX.Element | null {
-  if (props.entityListType !== "projects") return null;
-  if (!props.browserURL || !props.data) return null;
-  const jsonLd = buildHcaProjectJsonLd(
-    props.data as ProjectsResponse,
-    props.browserURL
-  );
-  return jsonLd ? <JsonLd jsonLd={jsonLd} /> : null;
-}
 
 /**
  * Returns the override for the given entity ID.
@@ -546,4 +528,22 @@ async function processEntityProps(
   if (entityResponse) {
     props.data = entityResponse;
   }
+}
+
+/**
+ * Renders the HCA project JSON-LD when the page is a project detail route with
+ * data and a browser URL available. Returns null otherwise.
+ * @param props - Entity detail page props.
+ * @returns JsonLd element, or null when the page can't be described.
+ */
+function renderHcaProjectJsonLd(
+  props: EntityDetailPageProps
+): JSX.Element | null {
+  if (props.entityListType !== "projects") return null;
+  if (!props.browserURL || !props.data) return null;
+  const jsonLd = buildHcaProjectJsonLd(
+    props.data as ProjectsResponse,
+    props.browserURL
+  );
+  return jsonLd ? <JsonLd jsonLd={jsonLd} /> : null;
 }
